Move logic into a ContentRequest class
