+++ src/Controller/Async/Records.php
@@ -76,7 +76,7 @@ class Records extends AsyncBase
 // sprintf('Attempt to modify invalid ContentType: %s', $contentTypeSlug);
                 continue;
             } else {
-                $this->modifyContentType($contentTypeSlug, $recordIds);
+                $this->app['storage.request.modify']->action($contentTypeSlug, $recordIds);
             }
         }
 
@@ -99,142 +99,5 @@ class Records extends AsyncBase
         ];
 
         return $this->render('@bolt/overview/overview.twig', $context);
-    }
-
-    /**
-     * Modify an individual ContentType's records.
-     *
-     * @param string $contentTypeSlug
-     * @param array  $recordIds
-     */
-    protected function modifyContentType($contentTypeSlug, array $recordIds)
-    {
-        foreach ($recordIds as $recordId => $actionData) {
-            if ($actionData === null) {
-                continue;
-            }
-
-            $repo = $this->getRepository($contentTypeSlug);
-            foreach ($actionData as $action => $fieldData) {
-                if (!$entity = $repo->find($recordId)) {
-                    continue;
-                }
-                $this->modifyContentTypeRecord($action, $repo, $entity, $fieldData);
-            }
-        }
-    }
-
-    /**
-     * Perform modification action(s) on a ContentType record.
-     *
-     * @param string     $action
-     * @param Repository $repo
-     * @param Content    $entity
-     * @param array      $fieldData
-     */
-    protected function modifyContentTypeRecord($action, $repo, $entity, $fieldData)
-    {
-        if ($action === 'delete') {
-            $this->deleteRecord($repo, $entity);
-        } elseif ($action === 'modify') {
-            $this->modifyRecord($entity, $fieldData);
-
-            if ($entity->_modified === true) {
-                if ($repo->save($entity)) {
-// $this->flashes()->info(Trans::__("Content '%title%' has been changed to '%newStatus%'", ['%title%' => $title, '%newStatus%' => $newStatus]));
-                } else {
-// $this->flashes()->info(Trans::__("Content '%title%' could not be modified.", ['%title%' => $title]));
-                }
-            }
-        }
-    }
-
-    /**
-     * Execute the deletion of a record.
-     *
-     * @param Repository $repo
-     * @param Content    $entity
-     */
-    protected function deleteRecord($repo, $entity)
-    {
-        $recordId = $entity->getId();
-        $contentTypeSlug = (string) $entity->getContenttype();
-        if (!$this->isAllowed("contenttype:$contentTypeSlug:delete:$recordId")) {
-            return;
-        }
-        $repo->delete($entity);
-    }
-
-    /**
-     * Modify a record's value(s).
-     *
-     * @param Content $entity
-     * @param array   $fieldData
-     */
-    protected function modifyRecord($entity, array $fieldData)
-    {
-        foreach ($fieldData as $field => $value) {
-            if (strtolower($field) === 'status') {
-                $this->transistionRecordStatus($entity, $value);
-            } elseif (strtolower($field) === 'ownerid') {
-                $this->transistionRecordOwner($entity, $value);
-            } else {
-                $this->modifyRecordValue($entity, $field, $value);
-            }
-        }
-    }
-
-    /**
-     * Modify a record's value if permitted.
-     *
-     * @param Content $entity
-     * @param string  $field
-     * @param mixed   $value
-     */
-    protected function modifyRecordValue($entity, $field, $value)
-    {
-        $recordId = $entity->getId();
-        $contentTypeSlug = (string) $entity->getContenttype();
-        $canModify = $this->isAllowed("contenttype:$contentTypeSlug:edit:$recordId");
-        if (!$canModify) {
-            return;
-        }
-        $entity->$field = Input::cleanPostedData($value);
-        $entity->_modified = true;
-    }
-
-    /**
-     * Transition a record's status if permitted.
-     *
-     * @param Content $entity
-     * @param string  $newStatus
-     */
-    protected function transistionRecordStatus($entity, $newStatus)
-    {
-        $contentTypeSlug = (string) $entity->getContenttype();
-        $canTransition = $this->users()->isContentStatusTransitionAllowed($entity->getStatus(), $newStatus, $contentTypeSlug, $entity->getId());
-        if (!$canTransition) {
-            return;
-        }
-        $entity->setStatus($newStatus);
-        $entity->_modified = true;
-    }
-
-    /**
-     * Transition a record's owner if permitted.
-     *
-     * @param Content $entity
-     * @param integer $ownerId
-     */
-    protected function transistionRecordOwner($entity, $ownerId)
-    {
-        $recordId = $entity->getId();
-        $contentTypeSlug = (string) $entity->getContenttype();
-        $canChangeOwner = $this->isAllowed("contenttype:$contentTypeSlug:change-ownership:$recordId");
-        if (!$canChangeOwner) {
-            return;
-        }
-        $entity->setOwnerid($ownerId);
-        $entity->_modified = true;
     }
 }
